Move read_flash_4k() to legacy interface
